chore: migrate test function
