+++ src/utils/SpokeUtils.ts
@@ -1,12 +1,11 @@
 import assert from "assert";
 import { BigNumber, BytesLike, Contract, PopulatedTransaction, providers, utils as ethersUtils } from "ethers";
-import { CHAIN_IDs, ZERO_ADDRESS } from "../constants";
+import { ZERO_ADDRESS } from "../constants";
 import { Deposit, Fill, FillStatus, RelayData, SlowFillRequest } from "../interfaces";
 import { SpokePoolClient } from "../clients";
 import { chunk } from "./ArrayUtils";
 import { toBN } from "./BigNumberUtils";
 import { isDefined } from "./TypeGuards";
-import { getNetworkName } from "./NetworkUtils";
 
 type BlockTag = providers.BlockTag;
 
@@ -304,69 +303,4 @@ export async function fillStatusArray(
       ? status.toNumber()
       : undefined;
   });
-}
-
-/**
- * Find the block at which a fill was completed.
- * @todo After SpokePool upgrade, this function can be simplified to use the FillStatus enum.
- * @param spokePool SpokePool contract instance.
- * @param relayData Deposit information that is used to complete a fill.
- * @param lowBlockNumber The lower bound of the search. Must be bounded by SpokePool deployment.
- * @param highBlocknumber Optional upper bound for the search.
- * @returns The block number at which the relay was completed, or undefined.
- */
-export async function findFillBlock(
-  spokePool: Contract,
-  relayData: RelayData,
-  lowBlockNumber: number,
-  highBlockNumber?: number
-): Promise<number | undefined> {
-  const { provider } = spokePool;
-  highBlockNumber ??= await provider.getBlockNumber();
-  assert(highBlockNumber > lowBlockNumber, `Block numbers out of range (${lowBlockNumber} > ${highBlockNumber})`);
-
-  // In production the chainId returned from the provider matches 1:1 with the actual chainId. Querying the provider
-  // object saves an RPC query becasue the chainId is cached by StaticJsonRpcProvider instances. In hre, the SpokePool
-  // may be configured with a different chainId than what is returned by the provider.
-  // @todo Sub out actual chain IDs w/ CHAIN_IDs constants
-  const destinationChainId = Object.values(CHAIN_IDs).includes(relayData.originChainId)
-    ? (await provider.getNetwork()).chainId
-    : Number(await spokePool.chainId());
-  assert(
-    relayData.originChainId !== destinationChainId,
-    `Origin & destination chain IDs must not be equal (${destinationChainId})`
-  );
-
-  // Make sure the relay war completed within the block range supplied by the caller.
-  const [initialFillStatus, finalFillStatus] = (
-    await Promise.all([
-      relayFillStatus(spokePool, relayData, lowBlockNumber, destinationChainId),
-      relayFillStatus(spokePool, relayData, highBlockNumber, destinationChainId),
-    ])
-  ).map(Number);
-
-  if (finalFillStatus !== FillStatus.Filled) {
-    return undefined; // Wasn't filled within the specified block range.
-  }
-
-  // Was filled earlier than the specified lowBlock. This is an error by the caller.
-  if (initialFillStatus === FillStatus.Filled) {
-    const { depositId, originChainId } = relayData;
-    const [srcChain, dstChain] = [getNetworkName(originChainId), getNetworkName(destinationChainId)];
-    throw new Error(`${srcChain} deposit ${depositId} filled on ${dstChain} before block ${lowBlockNumber}`);
-  }
-
-  // Find the leftmost block where filledAmount equals the deposit amount.
-  do {
-    const midBlockNumber = Math.floor((highBlockNumber + lowBlockNumber) / 2);
-    const fillStatus = await relayFillStatus(spokePool, relayData, midBlockNumber, destinationChainId);
-
-    if (fillStatus === FillStatus.Filled) {
-      highBlockNumber = midBlockNumber;
-    } else {
-      lowBlockNumber = midBlockNumber + 1;
-    }
-  } while (lowBlockNumber < highBlockNumber);
-
-  return lowBlockNumber;
 }
